[기영]useModal.js파일 추가 - RootStackNavigation에서 모달관련 정의부분 분리
